clippy: the crates we write are linted in CI too — the engine was the only one, so lucivy_core, lucistore, luciole and sparse_vector had never been
Nine lints in lucistore and six in lucivy_core, all mechanical: io::Error::other,
is_some_and / is_ok_and, named types for two very complex signatures
(MemFiles, IterTerms), an explicit drop of the broadcast handle (what
RamDirectory already does), a dead "parse" alternative the arm above
already matches, a parameter only passed to its own recursion, a cast to
its own type, a doc list. The two deprecated serialized_term calls are
allowed with the reason: that key is ExportableStats' wire format.

+++ lucivy_core/src/sharded_handle.rs
@@ -2471,7 +2471,7 @@ impl ShardedHandle {
                         listed - opened,
                     )];
                 }
-                let docs = self.num_docs().max(1) as u64;
+                let docs = self.num_docs().max(1);
                 // What the same index says a comfortable corpus would be, from
                 // its own measured bytes per document rather than a guess.
                 let fits = ram_index_max() / (index_bytes / docs).max(1);
